+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 1. `activate-keypads` function.
 2. _erlang_ bindings.
 
+### Updated
+1. Replaced os.Rename with lib implementation for tmpfs support.
+
 
 ## [0.8.5](https://github.com/uhppoted/uhppoted-codegen/releases/tag/v0.8.5) - 2023-06-13